+++ Userland/PinkOS/programs/time.c
@@ -19,6 +19,6 @@ extern void syscall(uint64_t syscall, uint64_t arg1, uint64_t arg2, uint64_t arg
 
 void time_main(unsigned char * args){
     RTC_Time time;
-    syscall(GET_RTC_TIME_SYSCALL, &time, 0, 0, 0, 0);
+    syscall(GET_RTC_TIME_SYSCALL, (uint64_t) &time, 0, 0, 0, 0);
     printf("Son las %d:%d:%d\n", (int)time.hours, (int)time.minutes, (int)time.seconds);
 }
